Refactor indexing to load with zeitwerk
Rails 7.2 required the zeitwerk path loader. Read more here:
https://guides.rubyonrails.org/classic_to_zeitwerk_howto.html

This means we now have to follow strict conventions about the location
and naming of files. This PR refactors our Traject indexing to conform
to those expectations.

Use bin/rails zeitwerk:check if you ever need to check zeitwerk loading
