+++ app/lib/dspace_indexer.rb
@@ -22,7 +22,7 @@ class DspaceIndexer
   end
 
   def research_data_config_path
-    pathname = ::Rails.root.join('lib', 'traject', "dataspace_research_data_config.rb")
+    pathname = ::Rails.root.join('config', 'traject', "dataspace_research_data_config.rb")
     pathname.to_s
   end
