+++ app/lib/indexing.rb
@@ -1,0 +1,5 @@
+module Indexing
+  require 'indexing/solr_cloud_helper'
+  require 'indexing/import_helper'
+  require 'indexing/domain'
+end
