+++ lib/tasks/index.rake
@@ -1,14 +1,11 @@
 # frozen_string_literal: true
 
-# Require ImportHelper so we can get the solr leader
-require ::Rails.root.join('lib', 'traject', 'solr_cloud_helper.rb')
-
 namespace :index do
   desc 'CRON JOB Re-index all research data'
   task research_data: :environment do
     Rails.logger.info "Indexing: Research Data indexing started"
-    SolrCloudHelper.create_collection_writer
-    Rails.logger.info "Indexing: Created a new collection for writing: #{SolrCloudHelper.collection_writer_url}"
+    Indexing::SolrCloudHelper.create_collection_writer
+    Rails.logger.info "Indexing: Created a new collection for writing: #{Indexing::SolrCloudHelper.collection_writer_url}"
 
     Rails.logger.info "Indexing: Fetching PDC Describe records"
     Rake::Task['index:pdc_describe_research_data'].invoke
@@ -16,15 +13,15 @@ namespace :index do
     Rake::Task['index:dspace_research_data'].invoke
     Rails.logger.info "Indexing: Fetching completed"
 
-    SolrCloudHelper.update_solr_alias!
-    Rails.logger.info "Indexing: Updated Solr to read from the new collection: #{SolrCloudHelper.alias_url} -> #{SolrCloudHelper.collection_reader_url}"
+    Indexing::SolrCloudHelper.update_solr_alias!
+    Rails.logger.info "Indexing: Updated Solr to read from the new collection: #{Indexing::SolrCloudHelper.alias_url} -> #{Indexing::SolrCloudHelper.collection_reader_url}"
   end
 
   desc 'Index all DSpace research data collections'
   task dspace_research_data: :environment do
     Rails.logger.info "Indexing: Harvesting and indexing DataSpace research data collections started"
     DspaceResearchDataHarvester.harvest(false)
-    SolrCloudHelper.collection_writer_commit!
+    Indexing::SolrCloudHelper.collection_writer_commit!
     Rails.logger.info "Indexing: Harvesting and indexing DataSpace research data collections completed"
   end
 
@@ -32,7 +29,7 @@ namespace :index do
   task pdc_describe_research_data: :environment do
     Rails.logger.info "Indexing: Harvesting and indexing PDC Describe data started"
     DescribeIndexer.new.index
-    SolrCloudHelper.collection_writer_commit!
+    Indexing::SolrCloudHelper.collection_writer_commit!
     Rails.logger.info "Indexing: Harvesting and indexing PDC Describe data completed"
   end
 
@@ -52,14 +49,14 @@ namespace :index do
 
   desc 'Prints to console the current Solr URLs and how they are configured'
   task print_solr_urls: :environment do
-    puts "Solr alias.: #{SolrCloudHelper.alias_url}"
-    puts "Solr reader: #{SolrCloudHelper.collection_reader_url}"
-    puts "Solr writer: #{SolrCloudHelper.collection_writer_url}"
+    puts "Solr alias.: #{Indexing::SolrCloudHelper.alias_url}"
+    puts "Solr reader: #{Indexing::SolrCloudHelper.collection_reader_url}"
+    puts "Solr writer: #{Indexing::SolrCloudHelper.collection_writer_url}"
   end
 
   desc 'Updates the Solr alias to point to the most up to date collection'
   task update_solr_alias: :environment do
-    puts "Solr updated: #{SolrCloudHelper.update_solr_alias!}"
+    puts "Solr updated: #{Indexing::SolrCloudHelper.update_solr_alias!}"
     Rake::Task['index:print_solr_urls'].invoke
   end
 
@@ -73,6 +70,6 @@ namespace :index do
       fixture_json = File.read(fixture_path)
       indexer.index_one(fixture_json)
     end
-    SolrCloudHelper.collection_writer_commit!
+    Indexing::SolrCloudHelper.collection_writer_commit!
   end
 end
