--- spec/fixtures/migration/dspace_migration.csv
@@ -1,2 +1,0 @@
-parent_community,community,collection_name,title,handle,ark_url,doi,curator,redescribed,pdc_describe_id,data_migrated
-Princeton Plasma Physics Laboratory,Advanced Projects,Other Projects,"Chatbots as social companions: Perceiving consciousness, human likeness, and social health benefits in machines",88435/dsp01t435gh23t,https://dataspace.princeton.edu/handle/88435/dsp01t435gh23t,"","","","",""
