+++ spec/system/dspace_indexing_spec.rb
@@ -8,7 +8,7 @@ describe 'DataSpace research data indexing', type: :system do
   end
   let(:indexer) do
     Traject::Indexer::NokogiriIndexer.new.tap do |i|
-      i.load_config_file(Rails.root.join('lib', 'traject', 'dataspace_research_data_config.rb'))
+      i.load_config_file(Rails.root.join('config', 'traject', 'dataspace_research_data_config.rb'))
     end
   end
   let(:dspace_xml) do
