+++ spec/system/pppl_indexing_spec.rb
@@ -18,7 +18,7 @@ describe 'PPPL research data indexing', type: :system do
   let(:output) { TrajectOutput.new }
   let(:indexer) do
     Traject::Indexer::NokogiriIndexer.new.tap do |i|
-      i.load_config_file(Rails.root.join('lib', 'traject', 'dataspace_research_data_config.rb'))
+      i.load_config_file(Rails.root.join('config', 'traject', 'dataspace_research_data_config.rb'))
     end
   end
   let(:dspace_xml) do
